Going through all step 1
