+++ shaders/base.fs
@@ -1,23 +1,28 @@
-#version 100
-
-precision mediump float;
+#version 330
 
 // Input vertex attributes (from vertex shader)
-varying vec2 fragTexCoord;
-varying vec4 fragColor;
+in vec2 fragTexCoord;
+in vec4 fragColor;
 
 // Input uniform values
 uniform sampler2D texture0;
 uniform vec4 colDiffuse;
+
+// Output fragment color
+out vec4 finalColor;
 
 // NOTE: Add your custom variables here
 
 void main()
 {
     // Texel color fetching from texture sampler
-    vec4 texelColor = texture2D(texture0, fragTexCoord);
+    vec4 texelColor = texture(texture0, fragTexCoord);
 
     // NOTE: Implement here your fragment shader code
 
-    gl_FragColor = texelColor*colDiffuse;
+    // final color is the color from the texture 
+    //    times the tint color (colDiffuse)
+    //    times the fragment color (interpolated vertex color)
+    finalColor = texelColor*colDiffuse*fragColor;
 }
+
