+++ src/GameConfig.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "Miscellaneous.h"
 #include "raylib.h"
 #include <string>
 #include <map>
@@ -10,17 +9,11 @@ public:
     std::map<std::string, std::string> texturePaths;
     std::map<std::string, std::string> shaderPaths;
     std::map<std::string, std::string> modelPaths;
-    int screenWidth;
-    int screenHeight;
-    int updateSpeed;
-    std::string windowTitle;
+    Vector2 screenSize;
     Color backgroundColor;
+    int updateRate;
+    float updateTime;
+    const char* windowTitle;
 
-    GameConfig(
-        int screenWidth,
-        int screenHeight,
-        int updateSpeed,
-        Color backgroundColor,
-        std::string windowTitle="Mr. AngryCube (DEV)"
-    );
+    GameConfig(Vector2 screenSize, Color backgroundColor);
 };
