+++ src/GameObject.cpp
@@ -1,56 +1,54 @@
 #include "GameObject.h"
-#include "raylib.h"
-#include "raymath.h"
-#include <typeinfo>
+#include "Utilities.h"
 
-
-GameObject::GameObject(Model& model, Shader& shader, Texture& texture)
-    : m_Model(model), m_Shader(shader), m_Texture(texture)
+GameObject::GameObject(Model* model, Material* material, Texture* texture)
+    : model(model), material(material), texture(texture)
 {
-    m_Material = LoadMaterialDefault();
-    m_Material.shader = m_Shader;
-
-    transform = MatrixIdentity();
-    SetMaterialTexture(&m_Material, MATERIAL_MAP_DIFFUSE, m_Texture);
+    objectId = Utilities::GenerateHash();
+    SetMaterialTexture(material, MATERIAL_MAP_DIFFUSE, *texture);
 }
 
 GameObject::~GameObject()
 {
-    UnloadModel(m_Model);
-    UnloadTexture(m_Texture);
-    UnloadMaterial(m_Material);
+    if (model != nullptr)
+    {
+        UnloadModel(*model);
+        model = nullptr;
+    }
+    if (material != nullptr)
+    {
+        UnloadMaterial(*material);
+        material = nullptr;
+    }
+    if (texture != nullptr)
+    {
+        UnloadTexture(*texture);
+        texture = nullptr;
+    }
 }
 
-void GameObject::Render()
-{
-    TraceLog(LOG_WARNING, "Subclasses should implement their own Render method");
-}
-
-void GameObject::Update(float deltaTime)
-{
-    TraceLog(LOG_WARNING, "Subclasses should implement their own Update method");
-}
-
-void GameObject::SetPosition(Vector3 position)
-{
-    TraceLog(LOG_INFO, "Subclasses should implement their own Update method");
-}
-
-Vector3 GameObject::GetPosition()
+Vector3
+GameObject::GetPosition()
 {
     return Vector3{transform.m12, transform.m13, transform.m14};
 }
 
-Vector3 GameObject::GetVelocity(float deltaTime, size_t begin, size_t end)
+void
+GameObject::SetPosition(Vector3 newPosition)
 {
-    if (m_Positions.size() == 0)
-    {
-        return Vector3();
-    }
+    transform.m12 = newPosition.x;
+    transform.m13 = newPosition.y;
+    transform.m14 = newPosition.z;
+}
 
-    Vector3 firstPosition = m_Positions.at(begin);
-    Vector3 lastPosition = m_Positions.at(std::min(end, m_Positions.size()) - 1);
-    Vector3 velocity = Vector3Scale(Vector3Subtract(lastPosition, firstPosition), 1.0f / deltaTime);
+void
+GameObject::Render()
+{
+    TraceLog(LOG_ERROR, "GameObject subclass should implement its own unique render method!");
+}
 
-    return velocity;
+void
+GameObject::Update(float deltaTime)
+{
+    TraceLog(LOG_ERROR, "GameObject subclass should implement its own unique update method!");
 }
