+++ src/MrAngryCube.h
@@ -1,34 +1,9 @@
-#pragma once
-#include "Game.h"
 #include "GameObject.h"
-#include "Miscellaneous.h"
-#include "raylib.h"
-#include "raymath.h"
-#include <vector>
-
 
 class MrAngryCube : public GameObject
 {
 public:
-    Vector3 nextRotationAxis = { 0.0f, 0.0f, 0.0f };
-    Vector3 rotationAxis     = { 0.0f, 0.0f, 0.0f };
-    Vector3 rotation         = { 0.0f, 0.0f, 0.0f };
-    float speed              = 0.0f;
-    bool canMove             = false; 
-
-    MrAngryCube(Model& model, Shader& shader, Texture& texture);
-    virtual ~MrAngryCube() = default;
+    MrAngryCube(Model* model, Material* material, Texture* texture);
+    ~MrAngryCube() = default;
     void Render() override;
-    void Update(float deltaTime) override;
-
-    bool IsFaceOnTheGround();
-    bool IsAngry();
-    bool IsAtQuarterRotation(bool ommitZero=true);
-    void WaitFor(float seconds);
-
-private:
-    float m_LastMovementCheckTime = 0.0f;
-    float m_Size = 0.0f;
-    float m_HalfSize = 0.0f;
-    float m_Hypotenuse = 0.0f;
 };
