+++ src/main.cpp
@@ -1,22 +1,9 @@
-#include "Enemy.h"
 #include "Game.h"
 #include "GameConfig.h"
-#include "GameObject.h"
-#include "Behaviours.h"
-#include "MrAngryCube.h"
-#include "SimpleGui.h"
-#include "raylib.h"
-#include "raymath.h"
-#include <vector>
+#include <memory>
 
-// TODO
-// #ifdef PLATFORM_WEB
-//     #define GLSL_VERSION 330
-// #else
-//     #define GLSL_VERSION 110
-// #endif
 
-int main(void)
+int main()
 {
     const char* wd = GetWorkingDirectory();
 
@@ -26,19 +13,11 @@ int main(void)
         TraceLog(LOG_INFO, "Setting everything for webplatform...");
     #endif
 
-    Game& game = Game::Get();
+    GameConfig gameConfig({1024, 768}, DARKGRAY);
+    gameConfig.updateRate = updateRate;
+    Game* gameInstance = &Game::Get();
+    gameInstance->Init(&gameConfig);
+    unsigned int returnCode = gameInstance->Run();
 
-    // GameConfig gameConfig(960*2, 540*2, updateRate, DARKGRAY);
-    GameConfig gameConfig(960, 540, updateRate, DARKGRAY);
-    gameConfig.modelPaths["macDefault"] = "./assets/models/mr_angry_cube_high_res.obj";
-    gameConfig.modelPaths["enemyDefault"] = "./assets/models/enemy.obj";
-    gameConfig.texturePaths["macDefault"] = "./assets/textures/texel_checker.png";
-    gameConfig.texturePaths["enemyDefault"] = "./assets/textures/concrete.png";
-    gameConfig.texturePaths["mainMenuBackground"] = "./assets/textures/MrCube-GDD.jpg";
-    gameConfig.shaderPaths["macDefault"] = "./assets/shaders/base.fs";
-
-    game.Init(&gameConfig);
-
-    int returnCode = game.Run();
     return returnCode;
 }
