+++ src___/Enemy.cpp
@@ -16,11 +16,38 @@ void Enemy::Render()
 
 void Enemy::Update(float deltaTime)
 {
+
+}
+
+void Enemy::Update_(float deltaTime, GameObject* other)
+{
     switch (state)
     {
     case GameObjectState::Alive:
-        transform = MatrixMultiply(MatrixRotateY(DEG2RAD * 1.0f), transform);
-        break;
+{
+        // Calculate the direction vector (XZ-plane only)
+        Vector3 direction = Vector3Subtract(other->GetPosition(), GetPosition());
+        direction.y = 0.0f; // Ignore the Y-axis for rotation
+        direction = Vector3Normalize(direction);
+
+        // Calculate the angle to rotate around the Y-axis
+        float angle = atan2(direction.x, direction.z); // Angle in radians
+
+        // Create a rotation matrix for the Y-axis
+        Matrix rotation = MatrixRotateY(angle);
+
+        Vector3 position = GetPosition();
+
+        // Apply the rotation to the transform matrix
+        transform = MatrixMultiply(rotation, MatrixIdentity());
+
+        transform.m12 = position.x;
+        transform.m13 = position.y;
+        transform.m14 = position.z;
+
+        // transform = MatrixMultiply(MatrixRotateY(DEG2RAD * 1.0f), transform);
+    }
+    break;
 
     case GameObjectState::Dead:
         break;
