+++ src___/GameConfig.cpp
@@ -1,0 +1,18 @@
+#include "GameConfig.h"
+
+GameConfig::GameConfig(
+    int screenWidth,
+    int screenHeight,
+    int updateSpeed,
+    Color backgroundColor,
+    std::string windowTitle
+)
+    :
+    screenWidth(screenWidth),
+    screenHeight(screenHeight),
+    updateSpeed(updateSpeed),
+    backgroundColor(backgroundColor),
+    windowTitle(windowTitle)
+{
+    Utilities::Log("Created game configuration.", "GAMECONFIG");
+}
