+++ src___/MovableGameObject.cpp
@@ -1,0 +1,18 @@
+#include "MovableGameObject.h"
+
+
+MovableGameObject::MovableGameObject(Model& model, Shader& shader, Texture& texture)
+    : GameObject(model, shader, texture)
+{
+    transform = MatrixScale(size.x, size.y, size.z);
+}
+
+void MovableGameObject::Render()
+{
+
+}
+
+void MovableGameObject::Update(float deltaTime)
+{
+
+}
